move togglemodebutton
