+++ client/src/components/Navbar.tsx
@@ -1,29 +1,27 @@
 import {
-  ActionIcon,
   Box,
   Burger,
   Button,
   Container,
-  createStyles,
   Group,
   Header,
   MediaQuery,
   Paper,
-  rem,
   Transition,
+  createStyles,
+  rem,
   useMantineColorScheme,
   useMantineTheme,
 } from '@mantine/core';
 import { useDisclosure } from '@mantine/hooks';
 import {
-  IconMoonStars,
   IconShoppingCart,
-  IconSunHigh,
-  IconUserShield,
+  IconUserShield
 } from '@tabler/icons-react';
 import { useEffect, useRef, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useShoppingCart } from '../contexts/ShoppingCartContext';
+import { ToggleColorButton } from './ToggleColorButton';
 
 const HEADER_HEIGHT = rem(70);
 
@@ -122,11 +120,6 @@ export function HeaderResponsive({ links }: HeaderResponsiveProps) {
   const [logoType, setLogoType] = useState('dark');
   const theme = useMantineTheme();
 
-  const handleToggle = () => {
-    toggleColorScheme();
-    setLogoType(colorScheme === 'dark' ? 'dark' : 'light');
-  };
-
   const logo =
     logoType === 'dark' ? (
       <img src="/assets/T101-logo.svg" alt="T101 logo" />
@@ -171,27 +164,6 @@ export function HeaderResponsive({ links }: HeaderResponsiveProps) {
     }
   }, [opened]);
 
-  function ToggleDarkAndLightMode() {
-    const { colorScheme } = useMantineColorScheme();
-    const dark = colorScheme === 'dark';
-
-    return (
-      <ActionIcon
-        variant="outline"
-        color={dark ? 'gray' : 'blue'}
-        onClick={handleToggle}
-        title="Toggle color scheme"
-        sx={{ marginRight: '1rem' }}
-      >
-        {dark ? (
-          <IconSunHigh size="1.3rem" stroke="1.6" />
-        ) : (
-          <IconMoonStars size="1.3rem" stroke="1.6" />
-        )}
-      </ActionIcon>
-    );
-  }
-
   function handleLinkClick() {
     setActive(links[0].link);
     window.scrollTo({
@@ -225,7 +197,7 @@ export function HeaderResponsive({ links }: HeaderResponsiveProps) {
           {items}
         </Group>
         <Group spacing={1}>
-          <ToggleDarkAndLightMode />
+          <ToggleColorButton />
           <Link to="/admin" data-cy="admin-link">
             <Button size="xs" variant="subtle" radius="xl">
               <IconUserShield size="1.8rem" stroke="1.3" />
